+++ evaluation/cli.py
@@ -9,9 +9,10 @@ import typer
 from rich.logging import RichHandler
 
 from evaluation import service_setup
-from evaluation.evaluate import get_models, run_evaluation_from_config
+from evaluation.evaluate import run_evaluation_from_config
 from evaluation.generate import generate_test_qa_answer, generate_test_qa_data
 from evaluation.red_teaming import run_red_teaming
+from evaluation.service_setup import get_models
 from evaluation.utils import load_config
 
 EVALUATION_DIR = Path(__file__).parent
@@ -51,30 +52,7 @@ def evaluate(
         exists=True,
         dir_okay=False,
         file_okay=True,
-        help=f"Path to the configuration JSON file. The name of the model to be evaluated should be specified in the JSON file. The available models that you can choose from are: {', '.join(get_models(get_model_url))}",
-        default=DEFAULT_CONFIG_PATH,
-    ),
-    numquestions: Optional[int] = typer.Option(
-        help="Number of questions to evaluate (defaults to all if not specified).",
-        default=None,
-        parser=int_or_none,
-    ),
-    targeturl: Optional[str] = typer.Option(
-        help="URL of the target service to evaluate (defaults to the value of the BACKEND_URI environment variable).",
-        default=None,
-        parser=str_or_none,
-    ),
-):
-    run_evaluation_from_config(EVALUATION_DIR, load_config(config), numquestions, targeturl)
-
-
-@app.command()
-def compare(
-    config: Path = typer.Option(
-        exists=True,
-        dir_okay=False,
-        file_okay=True,
-        help=f"Path to the configuration JSON file.The name of the models to be compared should be specified in the JSON file. The available models that you can choose from are: {', '.join(get_models(get_model_url))}",
+        help=f"Path to the configuration JSON file. The name of the model to be evaluated/compared should be specified in the JSON file. The available models that you can choose from are: {', '.join(get_models(get_model_url))}",
         default=DEFAULT_CONFIG_PATH,
     ),
     numquestions: Optional[int] = typer.Option(
@@ -142,7 +120,7 @@ def red_teaming(
         exists=True,
         dir_okay=False,
         file_okay=True,
-        help="Path to the configuration JSON file.",
+        help="Path to the configuration JSON file. The available models that can be used in the application and you can choose to compare are: {', '.join(get_models(get_model_url))}",
         default=DEFAULT_CONFIG_PATH,
     ),
     scorer_dir: Path = typer.Option(
@@ -182,46 +160,6 @@ def red_teaming(
             config=config,
             red_teaming_llm=red_team,
             prompt_target=target,
-            compare=False,
-            max_turns=max_turns,
-        )
-    )
-
-
-@app.command()
-def red_teaming_comparison(
-    config: Path = typer.Option(
-        exists=True,
-        dir_okay=False,
-        file_okay=True,
-        help=f"Path to the configuration JSON file. The available models that you can choose to compare are: {', '.join(get_models(get_model_url))}",
-        default=DEFAULT_CONFIG_PATH,
-    ),
-    scorer_dir: Path = typer.Option(
-        exists=True,
-        dir_okay=True,
-        file_okay=False,
-        help="Path to the directory where the scorer YAML files are stored.",
-        default=DEFAULT_SCORER_DIR,
-    ),
-    targeturl: Optional[str] = typer.Option(
-        help="URL of the target service to evaluate (defaults to the value of the BACKEND_URI environment variable).",
-        default=None,
-        parser=str_or_none,
-    ),
-    max_turns: int = typer.Option(default=3, help="The maximum number of turns to apply the attack strategy."),
-):
-    config = load_config(config)
-    red_team = service_setup.get_openai_target()
-    target = service_setup.get_app_target(config, targeturl)
-    asyncio.run(
-        run_red_teaming(
-            working_dir=EVALUATION_DIR,
-            scorer_dir=scorer_dir,
-            config=config,
-            red_teaming_llm=red_team,
-            prompt_target=target,
-            compare=True,
             max_turns=max_turns,
         )
     )
