+++ evaluation/config.json
@@ -3,7 +3,7 @@
     "results_dir": "results",
     "passing_rate": 3,
     "requested_metrics": ["gpt_groundedness", "gpt_relevance", "gpt_coherence", "gpt_similarity", "gpt_fluency", "answer_length", "latency", "f1_score"],
-    "models":["GPT 3.5 Turbo", "Mistral AI 7B", "Phi 3 Mini 4K"],
+    "models": ["GPT 3.5 Turbo", "Mistral AI 7B", "Phi 3 Mini 4K"],
     "max_workers": 1,
     "target_parameters": {
         "overrides": {
@@ -22,7 +22,7 @@
             ],
             "use_gpt4v": false,
             "gpt4v_input": "textAndImages",
-            "set_model":""
+            "set_model": ""
         }
     }
 }
