+++ evaluation/red_teaming.py
@@ -31,21 +31,17 @@ async def run_red_teaming(
     red_teaming_llm: PromptChatTarget,
     prompt_target: PromptChatTarget,
     max_turns: int,
-    compare: bool,
 ):
     """Run red teaming attack with provided scorers using Red Teaming Orchestrator."""
     prompt_target_list = []
-    if compare and isinstance(prompt_target, AppChatTarget):
+    if isinstance(prompt_target, AppChatTarget):
         compared_models = config.get("models")
         for compare_model in compared_models:
+            # Can't make a deep copy of the entire object because the internal DuckDBPyConnection object can't be copied directly.
             prompt_target_copy = copy.copy(prompt_target)
             prompt_target_copy.target_parameters = copy.deepcopy(prompt_target.target_parameters)
             prompt_target_copy.target_parameters["overrides"]["set_model"] = compare_model
             prompt_target_list.append(prompt_target_copy)
-    elif isinstance(prompt_target, AppChatTarget):
-        # Default use the first model in the config
-        prompt_target.target_parameters["overrides"]["set_model"] = config.get("models")[0]
-        prompt_target_list.append(prompt_target)
     else:
         prompt_target_list.append(prompt_target)
     logger.info("Running red teaming attack, with scorers from '%s'", scorer_dir)
@@ -61,7 +57,7 @@ async def run_red_teaming(
             model_name = prompt_target.__class__.__name__
         results_per_model = []
         for scorer_path in scorers:
-            logger.info("Runing red teaming with scorer YAML: %s", scorer_path)
+            logger.info("Running red teaming with scorer YAML: %s", scorer_path)
             try:
                 scorer_data = validate_scorer_yaml(scorer_path)
             except ValueError as e:
